Test Suite re-organized for a better nature package order.
