+++ src/test/java/com/mgz/afp/triplets/TripletTest.java
@@ -16,10 +16,9 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with Alpheus AFP Parser.  If not, see <http://www.gnu.org/licenses/>
 */
-package com.mgz.test;
+package com.mgz.afp.triplets;
 
 import com.mgz.afp.parser.TripletParser;
-import com.mgz.afp.triplets.Triplet;
 import com.mgz.afp.triplets.Triplet.TripletID;
 
 import org.junit.Test;
