+++ src/test/java/com/mgz/util/UtilBinaryDecodingTest.java
@@ -16,9 +16,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with Alpheus AFP Parser.  If not, see <http://www.gnu.org/licenses/>
 */
-package com.mgz.test;
-
-import com.mgz.util.UtilBinaryDecoding;
+package com.mgz.util;
 
 import org.junit.Test;
